+++ src/pages/home/ui/MainPage.tsx
@@ -1,4 +1,5 @@
-import { BottomBar, Header } from '../../../widgets';
+import BottomBar from '../../../widgets/main/ui/BottomBar';
+import Header from '../../../widgets/main/ui/Header';
 
 const MainPage = () => {
   return (
